// vscode/src/generated/diagram-product/DiagramProduct.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Completeness } from "./Completeness";
import type { DiagramDocument } from "./DiagramDocument";
import type { DiagramSource } from "./DiagramSource";
import type { Projection } from "./Projection";
import type { SelectedView } from "./SelectedView";
import type { SemanticReference } from "./SemanticReference";

export type DiagramProduct = { schemaVersion: 5, modelDigest: string, documents: Array<DiagramDocument>, sources: Array<DiagramSource>, references: Array<SemanticReference>, selectedView: SelectedView, completeness: Completeness, projection: Projection, };