// vscode/src/generated/diagram-product/SequenceScene.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SequenceMessage } from "./SequenceMessage";
import type { SequenceSceneKind } from "./SequenceSceneKind";

export type SequenceScene = { kind: SequenceSceneKind, lifelines: Array<number>, messages: Array<SequenceMessage>, };